Move responsibility from SQSMessageHandler to ResqueMessageHandler.
