+++ spec/sqs_message_handler_spec.rb
@@ -15,7 +15,7 @@ describe SQSMessageHandler do
   it "will log a 'young' message but not try to process it" do
     fake_message = double(:body => JSON.generate({}), message_attributes: {}, :attributes => {"SentTimestamp" => Time.now.to_i.to_s})
     message_handler = SQSMessageHandler.new({message: fake_message, settings: {'minimum_message_age_seconds' => "300"}})
-    expect_any_instance_of(NyplLogFormatter).to receive(:debug).with("Message '{}' is not old enough to process. It can be processed in 300 seconds")
+    expect(message_handler.instance_variable_get('@logger')).to receive(:debug).with("Message '{}' is not old enough to process. It can be processed in 300 seconds")
     message_handler.handle
   end
 
@@ -28,7 +28,7 @@ describe SQSMessageHandler do
     it "will log an error & delete the message" do
       fake_sqs_client = double(:delete_message)
       message_handler = SQSMessageHandler.new({sqs_client: fake_sqs_client, message: @fake_message, settings: {'sqs_queue_url' => 'http://example.com', 'minimum_message_age_seconds' => "300"}})
-      expect_any_instance_of(NyplLogFormatter).to receive(:error).with("Message '{\"action\":\"iamsupported\"}' contains an unsupported action")
+      expect(message_handler.instance_variable_get('@logger')).to receive(:error).with("Message '{\"action\":\"iamsupported\"}' contains an unsupported action")
       expect(fake_sqs_client).to receive(:delete_message).with(queue_url: 'http://example.com', receipt_handle: "some-id")
 
       message_handler.handle
